+++ jest.config.js
@@ -4,6 +4,7 @@ module.exports = {
     "^.+\\.vue$": "@vue/vue3-jest",
   },
   moduleFileExtensions: ["js", "ts", "json", "vue"],
+  testPathIgnorePatterns: ["/node_modules/", "/e2e/"],
   testEnvironment: "jsdom",
   testEnvironmentOptions: {
     customExportConditions: ["node", "node-addons"],